requirements.txt
langchain
langchain-google-genai
langgraph
google-generativeai
python-dotenv
langchain_community
langchain_chroma
pypdf